feat: Using Exceptions everywhere
I think we covered pretty much every case that could go wrong,
so that this auth_backend never crashes. It is now up to you
to implement a custom error_handler function and report some
exceptions to services like Sentry and forward others to simply
be displayed on the frontend.